requirements.txt
aiohttp==3.6.2
aiopg==1.0.0
aioredis==1.3.0
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.4.4
chardet==3.0.4
hiredis==1.0.1
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.6.1
psycopg2-binary==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
PyYAML==5.2
six==1.13.0
SQLAlchemy==1.3.11
typed-ast==1.4.0
wrapt==1.11.2
yarl==1.4.1
